Revert "Merge pull request #106 from alexanderwanyoike/feat/perform-card-28-cross-timeline-clipboard"
This reverts commit f2748b06278e1aef86950f236046b3c8775c0dcb, reversing
changes made to 61edb9d930556089b0860d88fa5e2f470825dfc3.
